MR B7: Channel adapter handler extraction
Decomposes telegram.py and discord.py into focused submodules.

- Telegram: handlers, attachments, outbound, approval
- Discord: approval_view, commands, history, attachments, outbound

Backward-compatible wrappers preserve all test contracts.
Metrics: telegram 735→310, discord 759→376.
